+++ packages/shared-types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tonightpass/shared-types
 
+## 2.0.5
+
+### Patch Changes
+
+- Add UpdateUserDto type
+
 ## 2.0.4
 
 ### Patch Changes